+++ smacc2_client_library/nav2z_client/custom_planners/forward_local_planner/src/forward_local_planner/forward_local_planner.cpp
@@ -544,7 +544,7 @@ geometry_msgs::msg::TwistStamped ForwardLocalPlanner::computeVelocityCommands(
   // check plan rejection
   bool aceptedplan = true;
 
-  unsigned int mx, my;
+  uint64_t mx, my;
 
   int i = 0;
   // RCLCPP_INFO_STREAM(nh_->get_logger(), "lplanner goal: " << finalgoalpose.pose.position);
